refactor: extract top budgets list
